+++ crates/client/db/src/lib.rs
@@ -1,5 +1,7 @@
 //! Madara database
 
+use alloy::primitives::private::alloy_rlp::MaxEncodedLenAssoc;
+use alloy::primitives::TxHash;
 use anyhow::{Context, Result};
 use bonsai_db::{BonsaiDb, DatabaseKeyMapping};
 use bonsai_trie::id::BasicId;
@@ -175,6 +177,7 @@ pub enum Column {
 
     L1Messaging,
     L1MessagingNonce,
+    L1MessagingHandlerTxHashes,
 
     /// Devnet: stores the private keys for the devnet predeployed contracts
     Devnet,
@@ -222,6 +225,7 @@ impl Column {
             BonsaiClassesLog,
             L1Messaging,
             L1MessagingNonce,
+            L1MessagingHandlerTxHashes,
             PendingContractToClassHashes,
             PendingContractToNonces,
             PendingContractStorage,
@@ -259,6 +263,7 @@ impl Column {
             ContractStorage => "contract_storage",
             L1Messaging => "l1_messaging",
             L1MessagingNonce => "l1_messaging_nonce",
+            L1MessagingHandlerTxHashes => "l1_messaging_handler_tx_hashes",
             PendingContractToClassHashes => "pending_contract_to_class_hashes",
             PendingContractToNonces => "pending_contract_to_nonces",
             PendingContractStorage => "pending_contract_storage",
@@ -285,6 +290,9 @@ impl Column {
                 opts.set_prefix_extractor(SliceTransform::create_fixed_prefix(
                     contract_db::CONTRACT_NONCES_PREFIX_EXTRACTOR,
                 ));
+            }
+            Column::L1MessagingHandlerTxHashes => {
+                opts.set_prefix_extractor(SliceTransform::create_fixed_prefix(TxHash::LEN));
             }
             _ => {}
         }