+++ crates/client/rpc/src/versions/v0_8_0/methods/read/mod.rs
@@ -1,2 +1,3 @@
 pub mod get_compiled_casm;
+pub mod get_messages_status;
 pub mod lib;